refactor(ui): align AdminGatedButton to canonical ButtonProps shape (#127)
Extract the inline AdminGatedButton from AppDetails.tsx into its own
component matching the canonical `extends ButtonProps` shape used by
cockpit-apt and cockpit-authelia-users. Forwarding all standard Button
props via spread replaces the restrictive hand-enumerated prop list
(variant union, label) that blocked legitimate PatternFly props.

Call sites switch from the `label` prop to `children`.
